Enunciate: Configuration for maven, and creation of a css
This allows to automate the creation of documentation, and defines a css.

However, we see a weird behavior with css. A change in CSS will only be affected when there is a change in the model. This is an enunciate bug.

Solution is to delete the target/ folder, so it will have to rebuild the entire documentation.

Change-Id: I764b54a152371ed7cc9cbccac9f7c343d367f3a0