ControllerTestConfiguration and TestCommissionController
ControllerTestConfiguration: Provides a centralized way to retrieve URLS for api tests.

TestCommissionController: Adds tests to check Commission API examples (on the website).

Current Problem: This tests work with the deployed configuration. they might FAIL if you change the data in the system.

Why I think this is good?

Because these are the examples AVAILABLE in the documentation. If the tests fail, you should ALSO correct the documentation.

We could change this to another model:
1. Get a standard "all" Page.
If it contains data:
2. get any id, name, etc... So we get data to create other tests.

Change-Id: I98bd289a1335099c915ac8eafdf2c4d749ab869f